requirements.txt
streamlit
python-docx
numpy
pandas
